+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## <next>
 
+## 1.10.0
+* Add `isClearable` prop to allow clearing of selected values.
+* Add option to hide group name input
+
 ## 1.9.1
 * Fix npm package content, include only necessary files.
 * Update docs to new format